Have the "TransformationStartsHere" section be its own component for reusability.
