+++ src/components/aboutus.js
@@ -3,8 +3,8 @@ import { Link } from "react-router-dom"; //Added Link import for the contact us 
 import "../style.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { motion } from "framer-motion";
+import TransformationStartsHere from "./transformationStartsHere.js";
 import Footer from "./footer.js";
-import { useNavigate } from "react-router-dom";
 import FounderImage from "../assets/founder.jpg";
 import MicrosoftLogo from "../assets/microsoft.png";
 import OracleLogo from "../assets/oracle.png";
@@ -14,13 +14,6 @@ import SalesLogo from "../assets/salesforce.png";
 
 const AboutUs = () => {
   const [inView, setInView] = useState(false);
-
-  const navigate = useNavigate();
-
-  const handleNavigation = (path) => {
-    navigate(path);
-    window.scrollTo(0, 0);
-  };
 
   const myRef = useRef();
   const chipsRef = useRef([]);
@@ -339,45 +332,8 @@ const AboutUs = () => {
         </div>
       </div>
 
-      <div className="footer-section">
-        <div className="footer-section-flex">
-          <div>
-            <h2>Transformation starts here</h2>
-            <br></br>
-            <p>At LVC Solutions, transformation starts where vision meets action. We’re not just a software company—we’re your dedicated partner in reimagining what's possible.<br />
-             Rooted in innovation and driven by real-world results, we empower businesses to cross the digital frontier with confidence and agility.<br />
-             Let's build tomorrow—starting now.</p>
-            <button
-              className="Connect-button"
-              onClick={() => {
-                handleNavigation("/contact");
-              }}
-            >
-              Connect With Us
-            </button>
-          </div>
-          <div className="divider"></div>
-          <div className="footer-section-flex-column">
-            <p>FIND OUT MORE</p>
-            <button
-              className="Connect-button-service"
-              onClick={() => {
-                handleNavigation("/services");
-              }}
-            >
-              Our Services
-            </button>
-            <button
-              className="Connect-button-contact "
-              onClick={() => {
-                handleNavigation("/contact");
-              }}
-            >
-              Contact Us
-            </button>
-          </div>
-        </div>
-      </div>
+      <TransformationStartsHere/>
+
       <Footer />
     </div>
   );
