+++ src/components/insight.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Link } from "react-router-dom";
 
@@ -10,17 +10,10 @@ import { Pagination, Navigation } from "swiper/modules";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faArrowRight } from "@fortawesome/free-solid-svg-icons";
 import "bootstrap/dist/css/bootstrap.min.css";
+import TransformationStartsHere from "./transformationStartsHere.js";
 import Footer from "./footer.js";
-import { useNavigate } from "react-router-dom";
 
 const Insight = () => {
-  const navigate = useNavigate();
-
-  const handleNavigation = (path) => {
-    navigate(path);
-    window.scrollTo(0, 0);
-  };
-
   const [showAllServices, setShowAllServices] = useState(false);
   const [inView, setInView] = useState(false);
   const toggleServices = () => {
@@ -331,45 +324,7 @@ const Insight = () => {
           </SwiperSlide>
         </Swiper>
       </div>
-      <div className="footer-section">
-        <div className="footer-section-flex">
-          <div>
-            <h2>Transformation starts here</h2>
-            <br></br>
-            <p>At LVC Solutions, transformation starts where vision meets action. We’re not just a software company—we’re your dedicated partner in reimagining what's possible.<br />
-             Rooted in innovation and driven by real-world results, we empower businesses to cross the digital frontier with confidence and agility.<br />
-             Let's build tomorrow—starting now.</p>
-            <button
-              className="Connect-button"
-              onClick={() => {
-                handleNavigation("/contact");
-              }}
-            >
-              Connect With Us
-            </button>
-          </div>
-          <div className="divider"></div>
-          <div className="footer-section-flex-column">
-            <p>FIND OUT MORE</p>
-            <button
-              className="Connect-button-service"
-              onClick={() => {
-                handleNavigation("/services");
-              }}
-            >
-              Our Services
-            </button>
-            <button
-              className="Connect-button-contact "
-              onClick={() => {
-                handleNavigation("/contact");
-              }}
-            >
-              Contact Us
-            </button>
-          </div>
-        </div>
-      </div>
+      <TransformationStartsHere/>
       <Footer />
     </main>
   );
